backend/requirements.txt
pytest==8.*
fastapi==0.*
uvicorn==0.*
python-multipart==0.*
httpx==0.*
python-dotenv==1.*
